+++ src/runtimes/kusama.rs
@@ -21,8 +21,8 @@
 
 use crate::config::CONFIG;
 use crate::crunch::{
-    get_from_seed, random_wait, try_fetch_stashes_from_remote_url, Crunch, NominatorsAmount,
-    ValidatorAmount, ValidatorIndex,
+    get_from_seed, random_wait, try_fetch_stashes_from_remote_url, Crunch,
+    NominatorsAmount, ValidatorAmount, ValidatorIndex,
 };
 use crate::errors::CrunchError;
 use crate::report::{
@@ -32,49 +32,54 @@ use crate::report::{
 use crate::stats;
 use async_recursion::async_recursion;
 use codec::Decode;
+use futures::StreamExt;
 use log::{debug, info, warn};
 use std::{
-    cmp, convert::TryFrom, convert::TryInto, fs, result::Result, str::FromStr, thread, time,
+    cmp, convert::TryFrom, convert::TryInto, fs, result::Result, str::FromStr, thread,
+    time,
 };
 use subxt::{
     sp_core::{sr25519, Pair as PairT},
     sp_runtime::AccountId32,
-    DefaultConfig, DefaultExtra, EventSubscription, PairSigner, RawEvent,
+    DefaultConfig, DefaultExtra, PairSigner,
 };
 
 #[subxt::subxt(
     runtime_metadata_path = "metadata/kusama_metadata.scale",
-    generated_type_derives = "Clone, Debug"
+    generated_type_derives = "Clone"
 )]
-mod api {}
+mod node_runtime {}
 
-pub type KusamaApi = api::RuntimeApi<DefaultConfig, DefaultExtra<DefaultConfig>>;
+use node_runtime::{
+    staking::events::EraPaid, staking::events::PayoutStarted, staking::events::Rewarded,
+    system::events::ExtrinsicFailed, utility::events::BatchCompleted,
+    utility::events::BatchInterrupted, utility::events::ItemCompleted,
+};
 
-type Call = api::runtime_types::kusama_runtime::Call;
-type StakingCall = api::runtime_types::pallet_staking::pallet::pallet::Call;
+pub type Api = node_runtime::RuntimeApi<DefaultConfig, DefaultExtra<DefaultConfig>>;
 
-pub async fn run_and_subscribe_era_paid_events(crunch: &Crunch) -> Result<(), CrunchError> {
+type Call = node_runtime::runtime_types::kusama_runtime::Call;
+type StakingCall = node_runtime::runtime_types::pallet_staking::pallet::pallet::Call;
+
+pub async fn run_and_subscribe_era_paid_events(
+    crunch: &Crunch,
+) -> Result<(), CrunchError> {
     info!("Inspect and `crunch` unclaimed payout rewards");
     // Run once before start subscription
     try_run_batch(&crunch, None).await?;
     info!("Subscribe 'EraPaid' on-chain finalized event");
     let client = crunch.client().clone();
-    let sub = client.rpc().subscribe_finalized_events().await?;
-    let decoder = client.events_decoder();
-    let mut sub = EventSubscription::<DefaultConfig>::new(sub, &decoder);
-    sub.filter_event::<api::staking::events::EraPaid>();
-    while let Some(result) = sub.next().await {
-        if let Ok(raw) = result {
-            match api::staking::events::EraPaid::decode(&mut &raw.data[..]) {
-                Ok(event) => {
-                    info!("Successfully decoded event {:?}", event);
-                    let wait: u64 = random_wait(120);
-                    info!("Waiting {} seconds before run batch", wait);
-                    thread::sleep(time::Duration::from_secs(wait));
-                    try_run_batch(&crunch, None).await?;
-                }
-                Err(e) => return Err(CrunchError::CodecError(e)),
-            }
+    let api = client.to_runtime_api::<Api>();
+    let mut sub = api.events().subscribe_finalized().await?;
+    while let Some(events) = sub.next().await {
+        let events = events?;
+
+        // Event --> staking::EraPaid
+        if let Some(_event) = events.find_first::<EraPaid>()? {
+            let wait: u64 = random_wait(120);
+            info!("Waiting {} seconds before run batch", wait);
+            thread::sleep(time::Duration::from_secs(wait));
+            try_run_batch(&crunch, None).await?;
         }
     }
     // If subscription has closed for some reason await and subscribe again
@@ -82,7 +87,10 @@ pub async fn run_and_subscribe_era_paid_events(crunch: &Crunch) -> Result<(), Cr
 }
 
 #[async_recursion]
-pub async fn try_run_batch(crunch: &Crunch, next_attempt: Option<u8>) -> Result<(), CrunchError> {
+pub async fn try_run_batch(
+    crunch: &Crunch,
+    next_attempt: Option<u8>,
+) -> Result<(), CrunchError> {
     // Skip run if it's the 3rd or more attempt
     let mut attempt = match next_attempt {
         Some(na) => {
@@ -96,7 +104,7 @@ pub async fn try_run_batch(crunch: &Crunch, next_attempt: Option<u8>) -> Result<
     };
     //
     let client = crunch.client();
-    let api = client.clone().to_runtime_api::<KusamaApi>();
+    let api = client.clone().to_runtime_api::<Api>();
     let properties = client.properties();
     let config = CONFIG.clone();
 
@@ -117,7 +125,8 @@ pub async fn try_run_batch(crunch: &Crunch, next_attempt: Option<u8>) -> Result<
     };
 
     // Get Token decimals
-    let token_decimals: u8 = if let Some(token_decimals) = properties.get("tokenDecimals") {
+    let token_decimals: u8 = if let Some(token_decimals) = properties.get("tokenDecimals")
+    {
         token_decimals
             .as_u64()
             .unwrap_or_default()
@@ -131,14 +140,14 @@ pub async fn try_run_batch(crunch: &Crunch, next_attempt: Option<u8>) -> Result<
     let network = Network {
         name: chain_name,
         active_era: active_era_index,
-        token_symbol: token_symbol,
-        token_decimals: token_decimals,
+        token_symbol,
+        token_decimals,
     };
     debug!("network {:?}", network);
 
     // Load seed account
-    let seed =
-        fs::read_to_string(config.seed_path).expect("Something went wrong reading the seed file");
+    let seed = fs::read_to_string(config.seed_path)
+        .expect("Something went wrong reading the seed file");
     let seed_account: sr25519::Pair = get_from_seed(&seed, None);
     let seed_account_signer =
         PairSigner::<DefaultConfig, DefaultExtra<DefaultConfig>, sr25519::Pair>::new(
@@ -160,7 +169,7 @@ pub async fn try_run_batch(crunch: &Crunch, next_attempt: Option<u8>) -> Result<
     let seed_account_info = api
         .storage()
         .system()
-        .account(seed_account_id, None)
+        .account(&seed_account_id, None)
         .await?;
     if seed_account_info.data.free <= (2 * ed) {
         signer
@@ -199,6 +208,11 @@ pub async fn try_run_batch(crunch: &Crunch, next_attempt: Option<u8>) -> Result<
         }
     }
     summary.total_validators = validators.len() as u32;
+    summary.total_validators_previous_era_already_claimed = validators
+        .iter()
+        .map(|v| v.claimed.contains(&(active_era_index - 1)) as u32)
+        .reduce(|a, b| a + b)
+        .unwrap_or_default();
 
     if calls_for_batch.len() > 0 {
         // TODO check batch call weight or maximum_calls [default: 8]
@@ -219,11 +233,13 @@ pub async fn try_run_batch(crunch: &Crunch, next_attempt: Option<u8>) -> Result<
                 let mut nominators_amount_value: NominatorsAmount = 0;
                 let mut nominators_quantity = 0;
 
-                let call_start_index: usize = (x * config.maximum_calls).try_into().unwrap();
+                let call_start_index: usize =
+                    (x * config.maximum_calls).try_into().unwrap();
                 let call_end_index: usize = if config.maximum_calls
                     > calls_for_batch[call_start_index..].len() as u32
                 {
-                    ((x * config.maximum_calls) + calls_for_batch[call_start_index..].len() as u32)
+                    ((x * config.maximum_calls)
+                        + calls_for_batch[call_start_index..].len() as u32)
                         .try_into()
                         .unwrap()
                 } else {
@@ -262,151 +278,119 @@ pub async fn try_run_batch(crunch: &Crunch, next_attempt: Option<u8>) -> Result<
                     0
                 };
 
-                let failed_event =
-                    tx_events.find_first_event::<api::system::events::ExtrinsicFailed>()?;
+                let failed_event = tx_events.find_first::<ExtrinsicFailed>()?;
 
                 if let Some(ev) = failed_event {
                     // TODO: repeat the batch call?  Or just log an error
-                    return Err(CrunchError::Other(format!("Extrinsic failed: {:?}", ev)));
+                    return Err(CrunchError::Other(format!(
+                        "Extrinsic failed: {:?}",
+                        ev
+                    )));
                 } else {
                     // Iterate over events to calculate respective reward amounts
-                    for event in tx_events.as_slice() {
+                    for event in tx_events.iter_raw() {
+                        let event = event?;
                         debug!("{:?}", event);
-                        match event {
-                            RawEvent {
-                                ref pallet,
-                                ref variant,
-                                data,
-                                ..
-                            } if pallet == "Staking" && variant == "PayoutStarted" => {
-                                // https://polkadot.js.org/docs/substrate/events#payoutstartedu32-accountid32
-                                // PayoutStarted(u32, AccountId32)
-                                // summary: The stakers' rewards are getting paid. [era_index, validator_stash]
-                                //
-                                let event_decoded =
-                                    api::staking::events::PayoutStarted::decode(&mut &data[..])?;
-                                debug!("{:?}", event_decoded);
-                                let validator_index_ref =
-                                    &mut validators.iter().position(|v| v.stash == event_decoded.1);
-                                era_index = event_decoded.0;
-                                validator_index = *validator_index_ref;
-                                validator_amount_value = 0;
-                                nominators_amount_value = 0;
-                                nominators_quantity = 0;
-                            }
-                            RawEvent {
-                                ref pallet,
-                                ref variant,
-                                data,
-                                ..
-                            } if pallet == "Staking" && variant == "Rewarded" => {
-                                // https://polkadot.js.org/docs/substrate/events#rewardedaccountid32-u128
-                                // Rewarded(AccountId32, u128)
-                                // summary: An account has been rewarded for their signed submission being finalized
-                                //
-                                let event_decoded =
-                                    api::staking::events::Rewarded::decode(&mut &data[..])?;
-                                debug!("{:?}", event_decoded);
-                                if let Some(i) = validator_index {
-                                    let validator = &mut validators[i];
-                                    if event_decoded.0 == validator.stash {
-                                        validator_amount_value = event_decoded.1;
-                                    } else {
-                                        nominators_amount_value += event_decoded.1;
-                                        nominators_quantity += 1;
-                                    }
-                                }
-                            }
-                            RawEvent {
-                                ref pallet,
-                                ref variant,
-                                ..
-                            } if pallet == "Utility" && variant == "ItemCompleted" => {
-                                // https://polkadot.js.org/docs/substrate/events#itemcompleted
-                                // summary: A single item within a Batch of dispatches has completed with no error.
-                                if let Some(i) = validator_index {
-                                    let validator = &mut validators[i];
-                                    // Add era to claimed vec
-                                    validator.claimed.push(era_index);
-                                    // Fetch stash points
-                                    let points = get_validator_points_info(
-                                        &crunch,
-                                        era_index,
-                                        validator.stash.clone(),
-                                    )
-                                    .await?;
-
-                                    let p = Payout {
-                                        block_number,
-                                        extrinsic: tx_events.extrinsic_hash(),
-                                        era_index,
-                                        validator_amount_value,
-                                        nominators_amount_value,
-                                        nominators_quantity,
-                                        points,
-                                    };
-                                    validator.payouts.push(p);
-                                    summary.calls_succeeded += 1;
-                                }
-                            }
-                            RawEvent {
-                                ref pallet,
-                                ref variant,
-                                ..
-                            } if pallet == "Utility" && variant == "BatchCompleted" => {
-                                // https://polkadot.js.org/docs/substrate/events#batchcompleted
-                                // summary: Batch of dispatches completed fully with no error.
-                                info!("Batch Completed for Era {}", network.active_era);
-                                attempt = None;
-                            }
-                            RawEvent {
-                                ref pallet,
-                                ref variant,
-                                data,
-                                ..
-                            } if pallet == "Utility" && variant == "BatchInterrupted" => {
-                                // https://polkadot.js.org/docs/substrate/events#batchinterruptedu32-spruntimedispatcherror
-                                // summary: Batch of dispatches did not complete fully. Index of first failing dispatch given, as well as the error.
-                                //
-                                // Fix: https://github.com/turboflakes/crunch/issues/4
-                                // Most likely the batch was interrupted because of an AlreadyClaimed era
-                                // BatchInterrupted { index: 0, error: Module { index: 6, error: 14 } }
-                                let event_decoded =
-                                    api::utility::events::BatchInterrupted::decode(&mut &data[..])?;
-                                warn!("{:?}", event_decoded);
-
-                                if let Call::Staking(call) = &calls_for_batch_clipped
-                                    [usize::try_from(event_decoded.index).unwrap()]
-                                {
-                                    match &call {
-                                        StakingCall::payout_stakers {
-                                            validator_stash, ..
-                                        } => {
-                                            warn!("validator_stash: {:?}", validator_stash);
-                                            let validator_index = &mut validators
-                                                .iter()
-                                                .position(|v| v.stash == *validator_stash);
-
-                                            if let Some(i) = *validator_index {
-                                                let validator = &mut validators[i];
-                                                // TODO: decode DispatchError to a readable format
-                                                validator
-                                                    .warnings
-                                                    .push("⚡ Batch interrupted ⚡".to_string());
-                                            }
-                                        }
-                                        _ => unreachable!(),
-                                    };
-                                }
-                                // Attempt to run one more time
-                                attempt = if let Some(na) = attempt {
-                                    Some(na + 1)
+                        if let Some(ev) = event.as_event::<PayoutStarted>()? {
+                            // https://polkadot.js.org/docs/substrate/events#payoutstartedu32-accountid32
+                            // PayoutStarted(u32, AccountId32)
+                            // summary: The stakers' rewards are getting paid. [era_index, validator_stash]
+                            //
+                            debug!("{:?}", ev);
+                            let validator_index_ref =
+                                &mut validators.iter().position(|v| v.stash == ev.1);
+                            era_index = ev.0;
+                            validator_index = *validator_index_ref;
+                            validator_amount_value = 0;
+                            nominators_amount_value = 0;
+                            nominators_quantity = 0;
+                        } else if let Some(ev) = event.as_event::<Rewarded>()? {
+                            // https://polkadot.js.org/docs/substrate/events#rewardedaccountid32-u128
+                            // Rewarded(AccountId32, u128)
+                            // summary: An account has been rewarded for their signed submission being finalized
+                            //
+                            debug!("{:?}", ev);
+                            if let Some(i) = validator_index {
+                                let validator = &mut validators[i];
+                                if ev.0 == validator.stash {
+                                    validator_amount_value = ev.1;
                                 } else {
-                                    Some(1)
+                                    nominators_amount_value += ev.1;
+                                    nominators_quantity += 1;
+                                }
+                            }
+                        } else if let Some(_ev) = event.as_event::<ItemCompleted>()? {
+                            // https://polkadot.js.org/docs/substrate/events#itemcompleted
+                            // summary: A single item within a Batch of dispatches has completed with no error.
+                            //
+                            if let Some(i) = validator_index {
+                                let validator = &mut validators[i];
+                                // Add era to claimed vec
+                                validator.claimed.push(era_index);
+                                // Fetch stash points
+                                let points = get_validator_points_info(
+                                    &crunch,
+                                    era_index,
+                                    &validator.stash,
+                                )
+                                .await?;
+
+                                let p = Payout {
+                                    block_number,
+                                    extrinsic: tx_events.extrinsic_hash(),
+                                    era_index,
+                                    validator_amount_value,
+                                    nominators_amount_value,
+                                    nominators_quantity,
+                                    points,
+                                };
+                                validator.payouts.push(p);
+                                summary.calls_succeeded += 1;
+                            }
+                        } else if let Some(_ev) = event.as_event::<BatchCompleted>()? {
+                            // https://polkadot.js.org/docs/substrate/events#batchcompleted
+                            // summary: Batch of dispatches completed fully with no error.
+                            info!("Batch Completed for Era {}", network.active_era);
+                            attempt = None;
+                        } else if let Some(ev) = event.as_event::<BatchInterrupted>()? {
+                            // https://polkadot.js.org/docs/substrate/events#batchinterruptedu32-spruntimedispatcherror
+                            // summary: Batch of dispatches did not complete fully. Index of first failing dispatch given, as well as the error.
+                            //
+                            // Fix: https://github.com/turboflakes/crunch/issues/4
+                            // Most likely the batch was interrupted because of an AlreadyClaimed era
+                            // BatchInterrupted { index: 0, error: Module { index: 6, error: 14 } }
+                            warn!("{:?}", ev);
+                            if let Call::Staking(call) = &calls_for_batch_clipped
+                                [usize::try_from(ev.index).unwrap()]
+                            {
+                                match &call {
+                                    StakingCall::payout_stakers {
+                                        validator_stash,
+                                        ..
+                                    } => {
+                                        warn!("validator_stash: {:?}", validator_stash);
+                                        let validator_index = &mut validators
+                                            .iter()
+                                            .position(|v| v.stash == *validator_stash);
+
+                                        if let Some(i) = *validator_index {
+                                            let validator = &mut validators[i];
+                                            // TODO: decode DispatchError to a readable format
+                                            validator.warnings.push(
+                                                "⚡ Batch interrupted ⚡".to_string(),
+                                            );
+                                        }
+                                    }
+                                    _ => unreachable!(),
                                 };
                             }
-                            _ => (),
-                        };
+                            // Attempt to run one more time
+                            attempt = if let Some(na) = attempt {
+                                Some(na + 1)
+                            } else {
+                                Some(1)
+                            };
+                        }
                     }
                 }
                 iteration = Some(x + 1);
@@ -442,7 +426,7 @@ async fn collect_validators_data(
     era_index: EraIndex,
 ) -> Result<Validators, CrunchError> {
     let client = crunch.client();
-    let api = client.clone().to_runtime_api::<KusamaApi>();
+    let api = client.clone().to_runtime_api::<Api>();
     let config = CONFIG.clone();
 
     // Get unclaimed eras for the stash addresses
@@ -459,7 +443,7 @@ async fn collect_validators_data(
         let stash = AccountId32::from_str(stash_str)?;
 
         // Check if stash has bonded controller
-        let controller = match api.storage().staking().bonded(stash.clone(), None).await? {
+        let controller = match api.storage().staking().bonded(&stash, None).await? {
             Some(controller) => controller,
             None => {
                 let mut v = Validator::new(stash.clone());
@@ -488,11 +472,8 @@ async fn collect_validators_data(
         let start_index = get_era_index_start(&crunch, era_index).await?;
 
         // Get staking info from ledger
-        if let Some(staking_ledger) = api
-            .storage()
-            .staking()
-            .ledger(controller.clone(), None)
-            .await?
+        if let Some(staking_ledger) =
+            api.storage().staking().ledger(&controller, None).await?
         {
             debug!(
                 "{} * claimed_rewards: {:?}",
@@ -503,6 +484,9 @@ async fn collect_validators_data(
             for e in (start_index..era_index).rev() {
                 // If reward was already claimed skip it
                 if staking_ledger.claimed_rewards.contains(&e) {
+                    if e == era_index - 1 {
+                        v.is_previous_era_already_claimed = true;
+                    }
                     v.claimed.push(e);
                     continue;
                 }
@@ -510,7 +494,7 @@ async fn collect_validators_data(
                 let exposure = api
                     .storage()
                     .staking()
-                    .eras_stakers(e, stash.clone(), None)
+                    .eras_stakers(&e, &stash, None)
                     .await?;
                 if exposure.total > 0 {
                     v.unclaimed.push(e)
@@ -528,7 +512,7 @@ async fn get_era_index_start(
     era_index: EraIndex,
 ) -> Result<EraIndex, CrunchError> {
     let client = crunch.client();
-    let api = client.clone().to_runtime_api::<KusamaApi>();
+    let api = client.clone().to_runtime_api::<Api>();
     let config = CONFIG.clone();
 
     let history_depth: u32 = api.storage().staking().history_depth(None).await?;
@@ -555,20 +539,20 @@ fn get_existential_deposit(crunch: &Crunch) -> Result<u128, CrunchError> {
 async fn get_validator_points_info(
     crunch: &Crunch,
     era_index: EraIndex,
-    stash: AccountId32,
+    stash: &AccountId32,
 ) -> Result<Points, CrunchError> {
     let client = crunch.client();
-    let api = client.clone().to_runtime_api::<KusamaApi>();
+    let api = client.clone().to_runtime_api::<Api>();
     // Get era reward points
     let era_reward_points = api
         .storage()
         .staking()
-        .eras_reward_points(era_index, None)
+        .eras_reward_points(&era_index, None)
         .await?;
     let stash_points = match era_reward_points
         .individual
         .iter()
-        .find(|(s, _)| *s == &stash)
+        .find(|(s, _)| *s == *stash)
     {
         Some((_, p)) => *p,
         None => 0,
@@ -600,14 +584,9 @@ async fn get_display_name(
     sub_account_name: Option<String>,
 ) -> Result<String, CrunchError> {
     let client = crunch.client();
-    let api = client.clone().to_runtime_api::<KusamaApi>();
+    let api = client.clone().to_runtime_api::<Api>();
 
-    match api
-        .storage()
-        .identity()
-        .identity_of(stash.clone(), None)
-        .await?
-    {
+    match api.storage().identity().identity_of(stash, None).await? {
         Some(identity) => {
             debug!("identity {:?}", identity);
             let parent = parse_identity_data(identity.info.display);
@@ -618,11 +597,8 @@ async fn get_display_name(
             Ok(name)
         }
         None => {
-            if let Some((parent_account, data)) = api
-                .storage()
-                .identity()
-                .super_of(stash.clone(), None)
-                .await?
+            if let Some((parent_account, data)) =
+                api.storage().identity().super_of(stash, None).await?
             {
                 let sub_account_name = parse_identity_data(data);
                 return get_display_name(
@@ -640,41 +616,109 @@ async fn get_display_name(
 }
 
 //
-fn parse_identity_data(data: api::runtime_types::pallet_identity::types::Data) -> String {
+fn parse_identity_data(
+    data: node_runtime::runtime_types::pallet_identity::types::Data,
+) -> String {
     match data {
-        api::runtime_types::pallet_identity::types::Data::Raw0(bytes) => str(bytes.to_vec()),
-        api::runtime_types::pallet_identity::types::Data::Raw1(bytes) => str(bytes.to_vec()),
-        api::runtime_types::pallet_identity::types::Data::Raw2(bytes) => str(bytes.to_vec()),
-        api::runtime_types::pallet_identity::types::Data::Raw3(bytes) => str(bytes.to_vec()),
-        api::runtime_types::pallet_identity::types::Data::Raw4(bytes) => str(bytes.to_vec()),
-        api::runtime_types::pallet_identity::types::Data::Raw5(bytes) => str(bytes.to_vec()),
-        api::runtime_types::pallet_identity::types::Data::Raw6(bytes) => str(bytes.to_vec()),
-        api::runtime_types::pallet_identity::types::Data::Raw7(bytes) => str(bytes.to_vec()),
-        api::runtime_types::pallet_identity::types::Data::Raw8(bytes) => str(bytes.to_vec()),
-        api::runtime_types::pallet_identity::types::Data::Raw9(bytes) => str(bytes.to_vec()),
-        api::runtime_types::pallet_identity::types::Data::Raw10(bytes) => str(bytes.to_vec()),
-        api::runtime_types::pallet_identity::types::Data::Raw11(bytes) => str(bytes.to_vec()),
-        api::runtime_types::pallet_identity::types::Data::Raw12(bytes) => str(bytes.to_vec()),
-        api::runtime_types::pallet_identity::types::Data::Raw13(bytes) => str(bytes.to_vec()),
-        api::runtime_types::pallet_identity::types::Data::Raw14(bytes) => str(bytes.to_vec()),
-        api::runtime_types::pallet_identity::types::Data::Raw15(bytes) => str(bytes.to_vec()),
-        api::runtime_types::pallet_identity::types::Data::Raw16(bytes) => str(bytes.to_vec()),
-        api::runtime_types::pallet_identity::types::Data::Raw17(bytes) => str(bytes.to_vec()),
-        api::runtime_types::pallet_identity::types::Data::Raw18(bytes) => str(bytes.to_vec()),
-        api::runtime_types::pallet_identity::types::Data::Raw19(bytes) => str(bytes.to_vec()),
-        api::runtime_types::pallet_identity::types::Data::Raw20(bytes) => str(bytes.to_vec()),
-        api::runtime_types::pallet_identity::types::Data::Raw21(bytes) => str(bytes.to_vec()),
-        api::runtime_types::pallet_identity::types::Data::Raw22(bytes) => str(bytes.to_vec()),
-        api::runtime_types::pallet_identity::types::Data::Raw23(bytes) => str(bytes.to_vec()),
-        api::runtime_types::pallet_identity::types::Data::Raw24(bytes) => str(bytes.to_vec()),
-        api::runtime_types::pallet_identity::types::Data::Raw25(bytes) => str(bytes.to_vec()),
-        api::runtime_types::pallet_identity::types::Data::Raw26(bytes) => str(bytes.to_vec()),
-        api::runtime_types::pallet_identity::types::Data::Raw27(bytes) => str(bytes.to_vec()),
-        api::runtime_types::pallet_identity::types::Data::Raw28(bytes) => str(bytes.to_vec()),
-        api::runtime_types::pallet_identity::types::Data::Raw29(bytes) => str(bytes.to_vec()),
-        api::runtime_types::pallet_identity::types::Data::Raw30(bytes) => str(bytes.to_vec()),
-        api::runtime_types::pallet_identity::types::Data::Raw31(bytes) => str(bytes.to_vec()),
-        api::runtime_types::pallet_identity::types::Data::Raw32(bytes) => str(bytes.to_vec()),
+        node_runtime::runtime_types::pallet_identity::types::Data::Raw0(bytes) => {
+            str(bytes.to_vec())
+        }
+        node_runtime::runtime_types::pallet_identity::types::Data::Raw1(bytes) => {
+            str(bytes.to_vec())
+        }
+        node_runtime::runtime_types::pallet_identity::types::Data::Raw2(bytes) => {
+            str(bytes.to_vec())
+        }
+        node_runtime::runtime_types::pallet_identity::types::Data::Raw3(bytes) => {
+            str(bytes.to_vec())
+        }
+        node_runtime::runtime_types::pallet_identity::types::Data::Raw4(bytes) => {
+            str(bytes.to_vec())
+        }
+        node_runtime::runtime_types::pallet_identity::types::Data::Raw5(bytes) => {
+            str(bytes.to_vec())
+        }
+        node_runtime::runtime_types::pallet_identity::types::Data::Raw6(bytes) => {
+            str(bytes.to_vec())
+        }
+        node_runtime::runtime_types::pallet_identity::types::Data::Raw7(bytes) => {
+            str(bytes.to_vec())
+        }
+        node_runtime::runtime_types::pallet_identity::types::Data::Raw8(bytes) => {
+            str(bytes.to_vec())
+        }
+        node_runtime::runtime_types::pallet_identity::types::Data::Raw9(bytes) => {
+            str(bytes.to_vec())
+        }
+        node_runtime::runtime_types::pallet_identity::types::Data::Raw10(bytes) => {
+            str(bytes.to_vec())
+        }
+        node_runtime::runtime_types::pallet_identity::types::Data::Raw11(bytes) => {
+            str(bytes.to_vec())
+        }
+        node_runtime::runtime_types::pallet_identity::types::Data::Raw12(bytes) => {
+            str(bytes.to_vec())
+        }
+        node_runtime::runtime_types::pallet_identity::types::Data::Raw13(bytes) => {
+            str(bytes.to_vec())
+        }
+        node_runtime::runtime_types::pallet_identity::types::Data::Raw14(bytes) => {
+            str(bytes.to_vec())
+        }
+        node_runtime::runtime_types::pallet_identity::types::Data::Raw15(bytes) => {
+            str(bytes.to_vec())
+        }
+        node_runtime::runtime_types::pallet_identity::types::Data::Raw16(bytes) => {
+            str(bytes.to_vec())
+        }
+        node_runtime::runtime_types::pallet_identity::types::Data::Raw17(bytes) => {
+            str(bytes.to_vec())
+        }
+        node_runtime::runtime_types::pallet_identity::types::Data::Raw18(bytes) => {
+            str(bytes.to_vec())
+        }
+        node_runtime::runtime_types::pallet_identity::types::Data::Raw19(bytes) => {
+            str(bytes.to_vec())
+        }
+        node_runtime::runtime_types::pallet_identity::types::Data::Raw20(bytes) => {
+            str(bytes.to_vec())
+        }
+        node_runtime::runtime_types::pallet_identity::types::Data::Raw21(bytes) => {
+            str(bytes.to_vec())
+        }
+        node_runtime::runtime_types::pallet_identity::types::Data::Raw22(bytes) => {
+            str(bytes.to_vec())
+        }
+        node_runtime::runtime_types::pallet_identity::types::Data::Raw23(bytes) => {
+            str(bytes.to_vec())
+        }
+        node_runtime::runtime_types::pallet_identity::types::Data::Raw24(bytes) => {
+            str(bytes.to_vec())
+        }
+        node_runtime::runtime_types::pallet_identity::types::Data::Raw25(bytes) => {
+            str(bytes.to_vec())
+        }
+        node_runtime::runtime_types::pallet_identity::types::Data::Raw26(bytes) => {
+            str(bytes.to_vec())
+        }
+        node_runtime::runtime_types::pallet_identity::types::Data::Raw27(bytes) => {
+            str(bytes.to_vec())
+        }
+        node_runtime::runtime_types::pallet_identity::types::Data::Raw28(bytes) => {
+            str(bytes.to_vec())
+        }
+        node_runtime::runtime_types::pallet_identity::types::Data::Raw29(bytes) => {
+            str(bytes.to_vec())
+        }
+        node_runtime::runtime_types::pallet_identity::types::Data::Raw30(bytes) => {
+            str(bytes.to_vec())
+        }
+        node_runtime::runtime_types::pallet_identity::types::Data::Raw31(bytes) => {
+            str(bytes.to_vec())
+        }
+        node_runtime::runtime_types::pallet_identity::types::Data::Raw32(bytes) => {
+            str(bytes.to_vec())
+        }
         _ => format!("???"),
     }
 }
@@ -685,7 +729,7 @@ fn str(bytes: Vec<u8>) -> String {
 
 pub async fn inspect(crunch: &Crunch) -> Result<(), CrunchError> {
     let client = crunch.client();
-    let api = client.clone().to_runtime_api::<KusamaApi>();
+    let api = client.clone().to_runtime_api::<Api>();
     let config = CONFIG.clone();
 
     info!("Inspect stashes -> {}", config.stashes.join(","));
@@ -702,12 +746,9 @@ pub async fn inspect(crunch: &Crunch) -> Result<(), CrunchError> {
         let mut unclaimed: Vec<u32> = Vec::new();
         let mut claimed: Vec<u32> = Vec::new();
 
-        if let Some(controller) = api.storage().staking().bonded(stash.clone(), None).await? {
-            if let Some(ledger_response) = api
-                .storage()
-                .staking()
-                .ledger(controller.clone(), None)
-                .await?
+        if let Some(controller) = api.storage().staking().bonded(&stash, None).await? {
+            if let Some(ledger_response) =
+                api.storage().staking().ledger(&controller, None).await?
             {
                 // Find unclaimed eras in previous 84 eras
                 for era_index in start_index..active_era_index {
@@ -720,7 +761,7 @@ pub async fn inspect(crunch: &Crunch) -> Result<(), CrunchError> {
                     let exposure = api
                         .storage()
                         .staking()
-                        .eras_stakers(era_index, stash.clone(), None)
+                        .eras_stakers(&era_index, &stash, None)
                         .await?;
                     if exposure.total > 0 {
                         unclaimed.push(era_index)